@@ -126,24 +126,24 @@
     "@electron-forge/maker-zip": "^6.0.0-beta.63",
     "@electron-forge/plugin-webpack": "^6.0.0-beta.63",
     "@electron-forge/publisher-github": "^6.0.0-beta.63",
-    "@pmmmwh/react-refresh-webpack-plugin": "^0.5.4",
-    "@types/d3": "^7.1.0",
+    "@pmmmwh/react-refresh-webpack-plugin": "^0.5.7",
+    "@types/d3": "^7.4.0",
     "@types/decompress": "^4.2.4",
-    "@types/jest": "^27.5.1",
+    "@types/jest": "^28.1.0",
     "@types/node-localstorage": "^1.3.0",
     "@types/os-utils": "^0.0.1",
-    "@types/react-dom": "^18.0.3",
+    "@types/react-dom": "^18.0.5",
     "@types/semver": "^7.3.9",
     "@types/uuid": "^8.3.4",
     "@types/yargs": "^17.0.10",
-    "@typescript-eslint/eslint-plugin": "^5.26.0",
-    "@typescript-eslint/parser": "^5.26.0",
-    "@vercel/webpack-asset-relocator-loader": "^1.7.0",
-    "babel-loader": "^8.2.2",
-    "concurrently": "^6.5.1",
+    "@typescript-eslint/eslint-plugin": "^5.27.0",
+    "@typescript-eslint/parser": "^5.27.0",
+    "@vercel/webpack-asset-relocator-loader": "^1.7.2",
+    "babel-loader": "^8.2.5",
+    "concurrently": "^7.2.1",
     "cross-env": "^7.0.3",
-    "css-loader": "^6.2.0",
-    "electron": "^18.2.0",
+    "css-loader": "^6.7.1",
+    "electron": "^19.0.2",
     "electron-installer-common": "^0.10.3",
     "eslint": "^8.16.0",
     "eslint-config-airbnb": "^19.0.4",
@@ -162,47 +162,47 @@
     "nodemon": "^2.0.16",
     "pre-commit": "^1.2.2",
     "prettier": "^2.6.2",
-    "react-refresh": "^0.11.0",
-    "semver-regex": ">=3.1.3",
-    "style-loader": "^3.2.1",
+    "react-refresh": "^0.13.0",
+    "semver-regex": ">=4.0.3",
+    "style-loader": "^3.3.1",
     "ts-jest": "^28.0.3",
     "ts-node": "^10.8.0",
     "typescript": "^4.7.2"
   },
   "dependencies": {
-    "@babel/plugin-transform-react-jsx": "^7.17.3",
-    "@chakra-ui/icons": "^1.0.15",
-    "@chakra-ui/react": "^1.8.8",
-    "@emotion/react": "^11.8.1",
+    "@babel/plugin-transform-react-jsx": "^7.17.12",
+    "@chakra-ui/icons": "^2.0.1",
+    "@chakra-ui/react": "^2.1.2",
+    "@emotion/react": "^11.9.0",
     "@emotion/styled": "^11.8.1",
-    "@fontsource/open-sans": "^4.5.1",
+    "@fontsource/open-sans": "^4.5.10",
     "@react-hook/window-size": "^3.0.7",
-    "@react-nano/use-event-source": "^0.11.0",
+    "@react-nano/use-event-source": "^0.12.0",
     "cross-fetch": "^3.1.5",
     "decompress": "^4.2.1",
-    "electron-log": "^4.4.1",
+    "electron-log": "^4.4.7",
     "electron-squirrel-startup": "^1.0.0",
-    "framer-motion": "^6.2.8",
+    "framer-motion": "^6.3.6",
     "html-to-image": "^1.9.0",
-    "image-js": "^0.33.0",
+    "image-js": "^0.34.1",
     "node-localstorage": "^2.2.1",
-    "nodejs-file-downloader": "^4.9.2",
+    "nodejs-file-downloader": "^4.9.3",
     "os-utils": "^0.0.14",
     "portfinder": "^1.0.28",
-    "re-resizable": "^6.9.5",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-flow-renderer": "^10.3.2",
-    "react-hotkeys-hook": "^3.4.4",
-    "react-icons": "^4.2.0",
+    "re-resizable": "^6.9.9",
+    "react": "^18.1.0",
+    "react-dom": "^18.1.0",
+    "react-flow-renderer": "^10.3.5",
+    "react-hotkeys-hook": "^3.4.6",
+    "react-icons": "^4.4.0",
     "react-markdown": "^8.0.3",
     "scheduler": "^0.22.0",
-    "semver": "^7.3.5",
-    "systeminformation": "^5.9.15",
-    "use-context-selector": "^1.3.10",
-    "use-debounce": "^7.0.1",
+    "semver": "^7.3.7",
+    "systeminformation": "^5.11.16",
+    "use-context-selector": "^1.4.0",
+    "use-debounce": "^8.0.1",
     "use-http": "^1.0.26",
     "uuid": "^8.3.2",
-    "yargs": "^17.5.0"
+    "yargs": "^17.5.1"
   }
 }
