@@ -1,6 +1,6 @@
 import { Center, ChakraProvider, Flex, Progress, Text, VStack } from '@chakra-ui/react';
 import { memo, useEffect, useState } from 'react';
-import ReactDOM from 'react-dom';
+import { createRoot } from 'react-dom/client';
 import { ipcRenderer } from '../common/safeIpc';
 import ChaiNNerLogo from './components/chaiNNerLogo';
 import './global.css';
@@ -107,8 +107,8 @@ const Splash = memo(() => {
                     >
                         <Center>
                             <Text
-                                isTruncated
                                 color="gray.500"
+                                textOverflow="ellipsis"
                             >
                                 {status}
                             </Text>
@@ -129,6 +129,8 @@ const Splash = memo(() => {
     );
 });
 
-ReactDOM.render(<Splash />, document.getElementById('root'));
+const container = document.getElementById('root');
+const root = createRoot(container!);
+root.render(<Splash />);
 
 export default Splash;
